@@ -14,7 +14,6 @@
     "selenium",
     "test"
   ],
-  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "wdio": "wdio run wdio.conf.ts",
@@ -23,15 +22,12 @@
     "wdio-docker": "DEBUG=false wdio run test/wdio.conf.ts && allure generate report allure-results --clean"
   },
   "devDependencies": {
-    "@wdio/allure-reporter": "^8.0.8",
-    "@wdio/cli": "^8.0.9",
-    "@wdio/globals": "^8.0.7",
-    "@wdio/jasmine-framework": "^8.0.9",
-    "@wdio/local-runner": "^8.0.9",
-    "@wdio/mocha-framework": "^8.0.9",
-    "@wdio/spec-reporter": "^8.0.8",
-    "cross-env": "^7.0.3",
-    "eslint": "^8.29.0",
+    "@wdio/cli": "^8.0.13",
+    "@wdio/globals": "^8.0.13",
+    "@wdio/jasmine-framework": "^8.0.13",
+    "@wdio/local-runner": "^8.0.13",
+    "@wdio/spec-reporter": "^8.0.13",
+    "eslint": "^8.30.0",
     "eslint-config-standard": "^17.0.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-node": "^11.1.0",
@@ -39,10 +35,8 @@
     "ts-node": "^10.9.1",
     "typescript": "^4.9.4",
     "wdio-chromedriver-service": "^8.0.1",
-    "webdriverio": "^8.0.7"
-  },
-  "dependencies": {
-    "expect-webdriverio": "^4.0.1",
-    "utility-types": "^3.10.0"
+    "webdriverio": "^8.0.13",
+    "@wdio/allure-reporter": "^8.0.13",
+    "cross-env": "^7.0.3"
   }
 }
